Rename identity creation flow to onboarding
